agents/requirements.txt
fastapi
langchain_anthropic
langchain_core
langgraph
confluent_kafka
uvicorn
python-dotenv
pymongo